refactor(webview): collapse three integration dialogs into one
